requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosmtpd==1.4.4.post2
pillow==10.4.0
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.0
alembic==1.13.0
email-validator==2.1.0
aiofiles==23.2.1
ffmpeg-python==0.2.0
psutil==7.1.0
